backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
aiofiles
psycopg2-binary
python-dotenv
